tests/wpt/web-platform-tests/app-history/navigate/state/updateCurrent/state-cross-document-location-navigate.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="i" src="/common/blank.html"></iframe>
<script>
async_test(t => {
  window.onload = t.step_func(() => {
    i.contentWindow.appHistory.updateCurrent({ state: { data: "value" } });
    assert_equals(i.contentWindow.appHistory.entries().length, 1);
    assert_equals(i.contentWindow.appHistory.current.getState().data, "value");

    i.contentWindow.location.href = "?1";
    i.onload = t.step_func_done(() => {
      assert_equals(i.contentWindow.appHistory.entries().length, 2);
      assert_equals(i.contentWindow.appHistory.current.index, 1);
      assert_equals(i.contentWindow.appHistory.current.getState(), undefined);
    });
  });
}, "appHistoryEntry.getState() behavior after cross-document location API navigation");
</script>
